wrap bare float literals in game_logic.py and constants.py with jnp.float32()
Prevents float64 promotion when jax_enable_x64=True by wrapping all bare
Python float literals in JAX comparison and arithmetic operations:
- 9 jax.random.uniform threshold comparisons (chest looting, sapling, mob AI)
- sleep damage multiplier (2.5)
- bow enchantment division (/ 2)
- BOSS_FIGHT_EXTRA_DAMAGE constant (0.5)

+++ craftax/craftax/constants.py
@@ -1,10 +1,12 @@
 import os
 import pathlib
 from enum import Enum
-import jax.numpy as jnp
+
 import imageio.v3 as iio
+import jax.numpy as jnp
 import numpy as np
 from PIL import Image
+
 from craftax.craftax.util.maths_utils import get_distance_map
 from craftax.environment_base.util import load_compressed_pickle, save_compressed_pickle
 
@@ -333,7 +335,7 @@ RANGED_MOB_TYPE_TO_PROJECTILE_TYPE_MAPPING = jnp.array(
 
 # GAME MECHANICS
 MONSTERS_KILLED_TO_CLEAR_LEVEL = 8
-BOSS_FIGHT_EXTRA_DAMAGE = 0.5
+BOSS_FIGHT_EXTRA_DAMAGE = jnp.float32(0.5)
 BOSS_FIGHT_SPAWN_TURNS = 7
 
 DIRECTIONS = jnp.concatenate(
